// frontend/src/app/core/api/model/timetable-slot-info-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TeacherInfo } from './teacher-info';
import { TimeSlotDefId } from './time-slot-def-id';
import { RoomInfo } from './room-info';
import { SubjectInfo } from './subject-info';


export interface TimetableSlotInfoResponse { 
    id?: number;
    dayOfWeek?: string;
    periodicity?: TimetableSlotInfoResponse.PeriodicityEnum;
    timeSlotDefinition?: TimeSlotDefId;
    subject?: SubjectInfo;
    teacher?: TeacherInfo;
    room?: RoomInfo;
}
export namespace TimetableSlotInfoResponse {
    export const PeriodicityEnum = {
        Normal: 'NORMAL',
        WeekA: 'WEEK_A',
        WeekB: 'WEEK_B'
    } as const;
    export type PeriodicityEnum = typeof PeriodicityEnum[keyof typeof PeriodicityEnum];
}